refactor: Rename `test/setupServerless`
