+++ integration-testing/wrapper.test.js
@@ -9,7 +9,7 @@ const log = require('log').get('test');
 const { ServerlessSDK } = require('@serverless/platform-client');
 const spawn = require('child-process-ext/spawn');
 const fixturesEngine = require('../test/fixtures');
-const setupServerless = require('../test/setupServerless');
+const setupServerless = require('../test/setup-serverless');
 
 describe('integration: wrapper', () => {
   let lambdaService;
