+++ lib/variables.test.js
@@ -4,7 +4,7 @@ const { expect } = require('chai');
 const path = require('path');
 const configUtils = require('@serverless/utils/config');
 const runServerless = require('../test/run-serverless');
-const setupServerless = require('../test/setupServerless');
+const setupServerless = require('../test/setup-serverless');
 
 const platformClientPath = require.resolve('@serverless/platform-client');
 const configUtilsPath = require.resolve('@serverless/utils/config');
