+++ scripts/setup-test-serverless.js
@@ -5,6 +5,6 @@
 require('essentials');
 require('log-node')();
 
-require('../test/setupServerless')({ shouldKeepServerlessDir: true }).then(({ root }) =>
+require('../test/setup-serverless')({ shouldKeepServerlessDir: true }).then(({ root }) =>
   process.stdout.write(`${root}\n`)
 );
