+++ test/run-serverless.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const setupServerless = require('./setupServerless');
+const setupServerless = require('./setup-serverless');
 
 module.exports = require('@serverless/test/setup-run-serverless-fixtures-engine')({
   fixturesDir: path.resolve(__dirname, './fixtures'),
